\name{Baby}
\docType{data}
\alias{Baby}
\alias{baby}
\alias{babyTable}

\title{Baby Dataset}

\description{
  Baby dataset 
}

\format{
  baby is a vector of data. babyTable is a data.frame
}

\details{
  This is a baby boom dataset summarizing the time of birth, sex, and
  birth weight for 44 babies born in one 24-hour period at a hospital in
  Brisbane, Australia. The full dataset can be found at \cite{A simple
  data set for demonstrating common distributions. Dunn, Peter
  K. Journal of Statistics Eduction. 7(3) 1999}
}

\keyword{datasets}





