\name{dbEmpLikeGOF-package}
\docType{package}
\alias{dbEmpLikeGOF-package}
\alias{EmpLikeGOF}
\title{Empirical Likelihood Goodness Of Fit Tests}
\description{
Package that has functions to perform empirical likelihood goodness of
fit tests for normality, uniformity, and distribution equality. 
}
\references{

  Jeffrey C. Miecznikowski, Albert Vexler, Lori A. Shepherd
  (2013). dbEmpLikeGOF: An R Package for Nonparametric Likelihood Ratio
  Tests for Goodness-of-Fit and Two-Sample Comparisons Based on Sample
  Entropy. Journal of Statistical Software, 54(3), 1-19.
  \url{http://www.jstatsoft.org/v54/i03/}
    
  Vexler A, Gurevich G, Empirical likelihood ratios applied to
  goodness-of-fit tests based on sample entropy.  Computational
  Statistics and Data Analysis 54(2010) 531-545.

  Gurevich G, Vexler A, A two-sample empirical likelihood ratio test
  based on samples entropy.  Statistics and Computing, 2011.
  
}
\seealso{\code{\link{dbEmpLikeGOF}}, \code{\link{returnCutoff}}}
\author{Jeffrey C. Mieznikowski, Lori A. Shepherd}
\keyword{package}
