\name{equation210}
\alias{equation210}
\title{Internal Function}
\description{internal function not called by user}
\usage{

equation210(x,
             y,
             m,
             i,
             j)

}

\arguments{
  \item{x}{vector of values to test}
  \item{y}{vector of values to test}
  \item{m}{value to loop over 1 to acceptable value}
  \item{i}{is either 1 or 2}
  \item{j}{loop of values 1 to sample size}
}

\details{
Function not called by user. Internal function only.
}

\value{comparison values}
\author{Jeffrey C. Miecznikowski, Lori A. Shepherd}
\keyword{internal}
