\name{Snow}
\docType{data}
\alias{Snow}
\alias{snow}

\title{Snow Dataset}

\description{
  Snowfall dataset 
}

\format{
  vector of values
}

\details{
  This file contains observations of the annual snowfall amounts in
  Buffalo, New York. 63 as observed from 1910/11 to 1972/73 as listed in
  \cite{The autoregressive method: a method of approximating and
  estimating positive functions. Carmichael, Jean-Pierre. DTIC
  Document. 1976}
  
}

\keyword{datasets}





