\name{Tbars}
\docType{data}
\alias{Tbars}
\alias{TBARS}
\alias{data}
\alias{x1}
\alias{y1}

\title{TBARS Dataset}

\description{
  TBARS dataset 
}

\format{
  data is a data.fram. x1 is a vector. y1 is a vector
}

\details{
  data from a study evaluating biomarkers related to atherosclerotic
  coronary heart disease 
}

\keyword{datasets}





