\name{testfun}
\alias{testfun}
\title{Internal Function for mapply}
\description{internal function not called by user}
\usage{
testfun(m,
        n,
        x.o)
}
\arguments{
  \item{m}{value to loop over}
  \item{n}{value to loop over}
  \item{x.o}{sorted list}
}
\details{  
  This function is used in a mapply within the db.emplike.gof function.
}
\value{Returns a vector of values}
\note{internal function not called by user}
\author{Jeffrey C. Miecznikowski, Lori A. Shepherd}
\seealso{\code{\link{dbEmpLikeGOF}}}
\keyword{internal}
